+++ front/components/editor/input_bar/MentionComponent.tsx
@@ -1,7 +1,7 @@
 import { NodeViewWrapper } from "@tiptap/react";
 import React from "react";
 
-import { MentionDisplay } from "@app/lib/mentions/ui/MentionDisplay";
+import { MentionDisplay } from "@app/components/mentions/MentionDisplay";
 import type { WorkspaceType } from "@app/types";
 
 interface MentionComponentProps {
