+++ front/components/mentions/MentionDisplay.tsx
@@ -15,11 +15,10 @@ import {
 } from "@dust-tt/sparkle";
 import React from "react";
 
+import { MentionDropdown } from "@app/components/mentions/MentionDropdown";
 import { useUser } from "@app/lib/swr/user";
 import type { WorkspaceType } from "@app/types";
 import type { RichMention } from "@app/types";
-
-import { MentionDropdown } from "./MentionDropdown";
 
 interface MentionDisplayProps {
   mention: RichMention;
