+++ front/lib/mentions/markdown/plugin.tsx
@@ -9,7 +9,7 @@ import { visit } from "unist-util-visit";
 
 import type { WorkspaceType } from "@app/types";
 
-import { MentionDisplay } from "../ui/MentionDisplay";
+import { MentionDisplay } from "@app/components/mentions/MentionDisplay";
 
 /*
  * Remark directive plugin for parsing mention directives.
